Make rust-mode optional
Do not depend on `rust-mode' because only a single function from that
package is required by a single function (which in turn is used by a
single command).  Do this so that users who use `rustic'[-mode] can
use `cargo' without having to install `rust-mode' just for that one
function.

Teach `cargo-process--get-current-test' to use the respective `rustic'
function.  If neither package installed then make it instruct the user
to install either one of them.